+++ src/linux/lconsole.c
@@ -245,7 +245,7 @@ int __al_linux_done_console (void)
 	    ret = write(STDERR_FILENO, msg, strlen(msg));
 	    if ((ret < 0) && (errno != EINTR))
 	       break;
-	 } while (ret < strlen(msg));
+	 } while (ret < (int)strlen(msg));
 
 	 __al_linux_got_text_message = FALSE;
       }
